move the logic of methods with filters from client to server side
