+++ Resources/Catalog/TSCatalogClient.py
@@ -4,9 +4,10 @@
 
 __RCSID__ = "$Id$"
 
+from DIRAC                                         import S_OK
+from DIRAC.Core.Utilities.List                     import breakListIntoChunks
 from DIRAC.Resources.Catalog.Utilities             import checkCatalogArguments
 from DIRAC.Resources.Catalog.FileCatalogClientBase import FileCatalogClientBase
-from DIRAC.TransformationSystem.Client.TransformationClient import TransformationClient
 
 class TSCatalogClient( FileCatalogClientBase ):
 
@@ -26,23 +27,28 @@ class TSCatalogClient( FileCatalogClientBase ):
 
   @checkCatalogArguments
   def addFile( self, lfns, force = False ):
-    """ Add file to the catalog
-    """
-    transClient = TransformationClient()
-    res = transClient.addFile( lfns, force )
-    return res
+    rpcClient = self._getRPC()
+    return rpcClient.addFile( lfns, force )
 
   @checkCatalogArguments
   def removeFile( self, lfns ):
-    transClient = TransformationClient()
-    res = transClient.removeFile( lfns )
-    return res
+    rpcClient = self._getRPC()
+    successful = {}
+    failed = {}
+    listOfLists = breakListIntoChunks( lfns, 100 )
+    for fList in listOfLists:
+      res = rpcClient.removeFile( fList )
+      if not res['OK']:
+        return res
+      successful.update( res['Value']['Successful'] )
+      failed.update( res['Value']['Failed'] )
+    resDict = {'Successful': successful, 'Failed':failed}
+    return S_OK( resDict )
 
   @checkCatalogArguments
   def setMetadata( self, path, metadatadict ):
     """ Set metadata parameter for the given path
         :return Successful/Failed dict.
     """
-    transClient = TransformationClient()
-    res = transClient.setMetadata( path, metadatadict )
-    return res
+    rpcClient = self._getRPC()
+    return rpcClient.setMetadata( path, metadatadict )
